pyproject.toml
[project]
name = "robyn-llamaindex-poc"
version = "0.1.0"
description = "A proof of concept application that combines Robyn with LlamaIndex for document analysis"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "robyn>=0.40.0",
    "llama-index>=0.10.0",
    "python-dotenv>=1.0.0",
    "openai>=1.12.0",
    "click>=8.1.7",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.26.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build]
include = [
    "main.py",
    "cli.py",
    "tests/**/*.py",
    "data/**/*",
    ".env.example",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 88
target-version = "py39" 
